fix(api): capabilities test - toleration seconds

+++ app/api/entity/capabilities_test.go
@@ -34,7 +34,7 @@ func getTestCapability() entity.Capabilities {
 				"operator":          "Equal",
 				"value":             "value1",
 				"effect":            "NoExecute",
-				"tolerationSeconds": int64(100),
+				"tolerationSeconds": 100,
 			},
 		},
 		Affinities: map[string]interface{}{},
